refactor: layoutに配置されていたコンポーネントをapperanceに移動 #19
